+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Add support for `is:banned` `not:banned` `include:banned` `exclude:banned`
   - Add support for `is:spoiled` `not:spoiled` `include:spoiled` `exclude:spoiled`
   - Add alias for `spoiled` with `previewed`
+  - Fix issue where id was not displayed in query messaage (#15)
 
 # 0.7.0